test organization
and a few field renames
