+++ addon_service/configured_storage_addon/serializers.py
@@ -3,6 +3,7 @@ from rest_framework_json_api.relations import ResourceRelatedField
 from rest_framework_json_api.utils import get_resource_type_from_model
 
 from addon_service.models import (
+    AuthorizedStorageAccount,
     ConfiguredStorageAddon,
     InternalResource,
 )
@@ -13,22 +14,22 @@ RESOURCE_NAME = get_resource_type_from_model(ConfiguredStorageAddon)
 
 class ConfiguredStorageAddonSerializer(serializers.HyperlinkedModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name=f"{RESOURCE_NAME}-detail")
-    authorized_storage_account = ResourceRelatedField(
-        queryset=ConfiguredStorageAddon.objects.all(),
+    base_account = ResourceRelatedField(
+        queryset=AuthorizedStorageAccount.objects.all(),
         many=False,
         related_link_view_name=f"{RESOURCE_NAME}-related",
     )
-    internal_resource = ResourceRelatedField(
+    authorized_resource = ResourceRelatedField(
         queryset=InternalResource.objects.all(),
         many=False,
         related_link_view_name=f"{RESOURCE_NAME}-related",
     )
 
     included_serializers = {
-        "authorized_storage_account": (
+        "base_account": (
             "addon_service.serializers.AuthorizedStorageAccountSerializer"
         ),
-        "internal_resource": "addon_service.serializers.InternalResourceSerializer",
+        "authorized_resource": "addon_service.serializers.InternalResourceSerializer",
     }
 
     class Meta:
@@ -36,6 +37,6 @@ class ConfiguredStorageAddonSerializer(serializers.HyperlinkedModelSerializer):
         fields = [
             "url",
             "root_folder",
-            "authorized_storage_account",
-            "internal_resource",
+            "base_account",
+            "authorized_resource",
         ]
