+++ addon_service/tests/test_by_type/test_internal_resource.py
@@ -3,63 +3,49 @@ import unittest
 from http import HTTPStatus
 
 from django.core.exceptions import ValidationError
-from django.db.models.query import QuerySet
 from django.test import TestCase
 from django.urls import reverse
 from rest_framework.test import APITestCase
 
 from addon_service import models as db
-from addon_service.internal_user.views import InternalUserViewSet
+from addon_service.internal_resource.views import InternalResourceViewSet
 from addon_service.tests import _factories
 from addon_service.tests._helpers import get_test_request
 
 
-# smoke-test api
-class TestInternalUserAPI(APITestCase):
+class TestInternalResourceAPI(APITestCase):
     @classmethod
     def setUpTestData(cls):
-        cls._user = _factories.InternalUserFactory()
+        cls._resource = _factories.InternalResourceFactory()
 
     @property
     def _detail_path(self):
-        return reverse("internal-users-detail", kwargs={"pk": self._user.pk})
+        return reverse("internal-resources-detail", kwargs={"pk": self._resource.pk})
 
     @property
     def _list_path(self):
-        return reverse("internal-users-list")
+        return reverse("internal-resources-list")
 
     @property
-    def _related_accounts_path(self):
+    def _related_configured_storage_addons_path(self):
         return reverse(
-            "internal-users-related",
+            "internal-resources-related",
             kwargs={
-                "pk": self._user.pk,
-                "related_field": "authorized_storage_accounts",
+                "pk": self._resource.pk,
+                "related_field": "configured_storage_addons",
             },
         )
 
     def test_get(self):
         _resp = self.client.get(self._detail_path)
         self.assertEqual(_resp.status_code, HTTPStatus.OK)
-        _content = json.loads(_resp.rendered_content)
-        self.assertEqual(
-            set(_content["data"]["attributes"].keys()),
-            {
-                "user_uri",
-            },
-        )
-        self.assertEqual(
-            set(_content["data"]["relationships"].keys()),
-            {
-                "authorized_storage_accounts",
-            },
-        )
+        self.assertEqual(_resp.data["resource_uri"], self._resource.resource_uri)
 
     def test_methods_not_allowed(self):
         _methods_not_allowed = {
             self._detail_path: {"patch", "put", "post"},
             # TODO: self._list_path: {'get', 'patch', 'put', 'post'},
-            self._related_accounts_path: {"patch", "put", "post"},
+            self._related_configured_storage_addons_path: {"patch", "put", "post"},
         }
         for _path, _methods in _methods_not_allowed.items():
             for _method in _methods:
@@ -70,61 +56,63 @@ class TestInternalUserAPI(APITestCase):
 
 
 # unit-test data model
-class TestInternalUserModel(TestCase):
+class TestInternalResourceModel(TestCase):
     @classmethod
     def setUpTestData(cls):
-        cls._user = _factories.InternalUserFactory()
+        cls._resource = _factories.InternalResourceFactory()
 
     def test_can_load(self):
-        _user_from_db = db.InternalUser.objects.get(id=self._user.id)
-        self.assertEqual(self._user.user_uri, _user_from_db.user_uri)
+        _resource_from_db = db.InternalResource.objects.get(id=self._resource.id)
+        self.assertEqual(self._resource.resource_uri, _resource_from_db.resource_uri)
 
-    def test_authorized_storage_accounts__empty(self):
-        _authed_storage_accounts_qs = self._user.authorized_storage_accounts
-        self.assertIsInstance(_authed_storage_accounts_qs, QuerySet)
-        self.assertEqual(list(_authed_storage_accounts_qs), [])
+    def test_configured_storage_addons__empty(self):
+        self.assertEqual(
+            list(self._resource.configured_storage_addons.all()),
+            [],
+        )
 
-    def test_authorized_storage_accounts__several(self):
+    def test_configured_storage_addons__several(self):
         _accounts = set(
-            _factories.AuthorizedStorageAccountFactory.create_batch(
+            _factories.ConfiguredStorageAddonFactory.create_batch(
                 size=3,
-                external_account__owner=self._user,
+                authorized_resource=self._resource,
             )
         )
-        _authed_storage_accounts_qs = self._user.authorized_storage_accounts
-        self.assertIsInstance(_authed_storage_accounts_qs, QuerySet)
-        self.assertEqual(set(_authed_storage_accounts_qs), _accounts)
+        self.assertEqual(
+            set(self._resource.configured_storage_addons.all()),
+            _accounts,
+        )
 
     def test_validation(self):
-        self._user.user_uri = "not a uri"
+        self._resource.resource_uri = "not a uri"
         with self.assertRaises(ValidationError):
-            self._user.clean_fields(exclude=["modified"])
+            self._resource.clean_fields(exclude=["modified"])
 
 
 # unit-test viewset (call the view with test requests)
-class TestInternalUserViewSet(TestCase):
+class TestInternalResourceViewSet(TestCase):
     @classmethod
     def setUpTestData(cls):
-        cls._user = _factories.InternalUserFactory()
-        cls._view = InternalUserViewSet.as_view({"get": "retrieve"})
+        cls._resource = _factories.InternalResourceFactory()
+        cls._view = InternalResourceViewSet.as_view({"get": "retrieve"})
 
     def test_get(self):
         _resp = self._view(
-            get_test_request(user=self._user),
-            pk=self._user.pk,
+            get_test_request(),
+            pk=self._resource.pk,
         )
         self.assertEqual(_resp.status_code, HTTPStatus.OK)
         _content = json.loads(_resp.rendered_content)
         self.assertEqual(
             set(_content["data"]["attributes"].keys()),
             {
-                "user_uri",
+                "resource_uri",
             },
         )
         self.assertEqual(
             set(_content["data"]["relationships"].keys()),
             {
-                "authorized_storage_accounts",
+                "configured_storage_addons",
             },
         )
 
@@ -143,34 +131,36 @@ class TestInternalUserViewSet(TestCase):
         self.assertEqual(_resp.status_code, HTTPStatus.FORBIDDEN)
 
 
-class TestInternalUserRelatedView(TestCase):
+class TestInternalResourceRelatedView(TestCase):
     @classmethod
     def setUpTestData(cls):
-        cls._user = _factories.InternalUserFactory()
-        cls._related_view = InternalUserViewSet.as_view({"get": "retrieve_related"})
+        cls._resource = _factories.InternalResourceFactory()
+        cls._related_view = InternalResourceViewSet.as_view(
+            {"get": "retrieve_related"},
+        )
 
     def test_get_related__empty(self):
         _resp = self._related_view(
-            get_test_request(user=self._user),
-            pk=self._user.pk,
-            related_field="authorized_storage_accounts",
+            get_test_request(),
+            pk=self._resource.pk,
+            related_field="configured_storage_addons",
         )
         self.assertEqual(_resp.status_code, HTTPStatus.OK)
         self.assertEqual(_resp.data, [])
 
     def test_get_related__several(self):
-        _accounts = _factories.AuthorizedStorageAccountFactory.create_batch(
+        _addons = _factories.ConfiguredStorageAddonFactory.create_batch(
             size=5,
-            external_account__owner=self._user,
+            authorized_resource=self._resource,
         )
         _resp = self._related_view(
-            get_test_request(user=self._user),
-            pk=self._user.pk,
-            related_field="authorized_storage_accounts",
+            get_test_request(),
+            pk=self._resource.pk,
+            related_field="configured_storage_addons",
         )
         self.assertEqual(_resp.status_code, HTTPStatus.OK)
         _content = json.loads(_resp.rendered_content)
         self.assertEqual(
             {_datum["id"] for _datum in _content["data"]},
-            {str(_account.pk) for _account in _accounts},
+            {str(_addon.pk) for _addon in _addons},
         )
